+++ services/src/main/java/dev/glabay/features/devices/CustomerDeviceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CustomerDeviceRepository extends JpaRepository<CustomerDevice, Integer> {
-    List<CustomerDevice> findByCustomerId(Integer customerId);
+    List<CustomerDevice> findByCustomerId(Long customerId);
 }
